requirements.txt
torch
torchvision
streamlit
gdown
Pillow
